+++ zenml/steps/trainer/pytorch_trainers/utils.py
@@ -137,21 +137,3 @@ def create_tf_dataset(file_pattern,
     dataset = dataset.map(_split_inputs_labels)
     dataset = dataset.prefetch(prefetch_buffer_size)
     return dataset
-
-
-def combine_batch_results(x):
-    result = {}
-    for batch in x:
-        for feature, values in batch.items():
-            if isinstance(values, torch.Tensor):
-                temp = torch.squeeze(values).detach().numpy()
-                values = np.reshape(temp, values.size())
-
-            if feature not in result:
-                result[feature] = values
-
-            else:
-                result[feature] = np.concatenate((result[feature],
-                                                  values), axis=0)
-
-    return result
